db sorgusu yapıldı.

+++ Domain/DomainService/Services/TransferOrderService.cs
@@ -22,12 +22,24 @@ namespace DomainService.Services
 
         public async Task<IList<TransferOrder>> GetAllAsync()
         {
-            return await _repository.GetAllAsync();
+            var transferOrders = await _repository.GetAllAsync();
+            if (transferOrders ==null || transferOrders.Count==0) {
+                throw new Exception("Not found any order");
+            }
+            else
+            {
+                return transferOrders;
+            }
         }
 
         public async Task<TransferOrder> GetByIdAsync(long id)
         {
-            return await _repository.GetByIdAsync(id);
+            var foundOrder=await _repository.GetByIdAsync(id);
+            if (foundOrder == null)
+            {
+                throw new Exception("Not found order");
+            }
+            else { return foundOrder; }
         }
 
         public async Task<IList<TransferOrder>> IncludeAsync(params string[] relations)
@@ -37,12 +49,35 @@ namespace DomainService.Services
 
         public async Task RemoveAsync(TransferOrder entity)
         {
-            await _repository.RemoveAsync(entity);
+            var foundOrder = await _repository.GetByIdAsync(entity.Id);
+            if (foundOrder == null)
+            {
+                throw new Exception("Not found");
+            }
+            else
+            {
+                await _repository.RemoveAsync(entity);
+            }
         }
 
         public async Task<long> UpdateAsync(TransferOrder entity)
         {
-            await _repository.UpdateAsync(entity);
+            var existingOrder = await GetByIdAsync(entity.Id);
+            if (existingOrder == null)
+            {
+                throw new Exception("Update failed");
+            }
+            else
+            {
+                existingOrder.StartingDate=entity.StartingDate;
+                existingOrder.StartingPosition=entity.StartingPosition;
+                existingOrder.EndPosition=entity.EndPosition;
+                existingOrder.Hostess=entity.Hostess;
+                existingOrder.Id=entity.Id;
+                existingOrder.AgentId=entity.AgentId;
+                existingOrder.IsActive=entity.IsActive;
+                existingOrder.VehicleId=entity.VehicleId;
+            }
             return entity.Id;
         }
